Resume_Chatter/requirements.txt
fastapi
aiofiles
PyPDF2
python-dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu